move to config.go
